Intelligently decide what to print in limited screen space
Function calls are often too large for a single line, messing up the menu.
That's probably primarily a TerminalMenus bug (and it should be fixed to
support multi-line display), but even so in extreme cases (*cough* Zygote),
function calls can be multiple terminal pages. This implements some logic
to always print function calls as a single line, falling back through the
following progression:
- Full argument types
- Only the head datatype, but not their parameters
- Only the number of arguments
- Nothing

E.g.:
```
In `Zygote._forward` select a call to descend into or ↩ to ascend. [q] to quit.
   %5 = invoke Zygote._forward(…,…,…,…)
   %18 = invoke getindex(::Tuple,::Int64)::Union{}
   %33 = invoke getindex(::Tuple,::Int64)::Union{}
   %65 = invoke getindex(::Tuple,::Int64)::Union{}
   %80 = invoke getindex(::Tuple,::Int64)::Union{}
   %270 = invoke getindex(::Tuple,::Int64)::Union{}
```