Add special handling for return_type
The idea is that the user will generally want to descend into
function that we're querying the return type of. However,
to differentiate this from a regular function call, I add a
special kind of callsite info to indicate this kind of call site
such that we may print it specially.