Preserve the Expr head when printing callsites (#113)
It's useful to be able to distinguish between compile-time and
run-time dispatch. Internally Julia does this with the `:invoke`/`:call`
distinction, but Cthulhu has historically used `:invoke` for both,
only using `:call` when multiple MethodInstances might apply.
Unfortunately, the multi-MI analysis does not take full account of
specialization, and so does not fully replicate the original pattern.

Here, we directly preserve the original head, which should make it
more robust to mistakes or changes in how Julia does things.