Minimal switch to AbstractInterpreter (#126)
Minimal switch to AbstractInterpreter

This is the minimal work required to swtich Cthulhu to AbstractInterpreter.
There are several things I would like to build on top of this:
    - Read callsites out of statment info rather than reconstructing it
    - Support for Opaque Closures
    - More consistent use of AbstractInterpreter throughout the API
    - Expose inference remarks

I would like to propose that we make a codebase break at Julia 1.6
for Cthulhu, moving the current master to a branch and requiring
Julia 1.7 on Cthulhu master. That gives us the rest of the 1.7 cycle
to update Cthulhu and figure out what the right interfaces between
Cthulhu and Base are.