WIP: Switch callsite detection to use stmt info
Rips out all the heuristics that are duplicating the analysis
that inference already does to figure out what is getting called.
This should be much more precise and general. We may want to
revisit the whole CallInfo setup in the future to maybe use the
stmt info more directly, but this should be good for the moment.

Depends on https://github.com/JuliaLang/julia/pull/39716 (so WIP
until that is merged).