fix #142
So the problem was kinda deeply nested than I expected.

The first thing to notice is that there are cases when we can't find a
cache in `CthulhuInterpreter.opt`.
AFAIU this is mainly because of limited frames; the native interpreter
doesn't cache them but there are still statement infos about them and
so the problem happens if a Cthulhu user tries to inspect it.

When trying to inspect uncached frames, the next problem is how to
handle them at pre-optimization phase. The `slottypes` of them would be
`nothing` at that point, and so we need to do something like
[what the external `OptimizationState` constructor
does](https://github.com/JuliaLang/julia/blob/878e1cd52ebf14e8666d8d54a54a95ccee4405d3/base/compiler/optimize.jl#L69-L71).
The another problem is the return types of those frames can be annotated
as `LimitedAccuracy` at that point, and it contains recursive frame
chain
and thus the printer can fail into infinite loop. The same issue happens
when Cthulhu tries to show `CodeInfo` object, which may also contain
`LimitedAccuracy`.

Now this commit addresses the issues outlined above:
- check if a callsite is cached or not, and wrap it into
`UncachedCallInfo` if not,
  and Cthulhu doesn't recurse into the `UncachedCallInfo` (or we can
even
  forcibly recurse into, but I decided not to do that, since it's really
  different behaviour from what the native interpreter did).
  Note that `UncachedCallInfo` won't happen when working on a
pre-optimization state,
  since Cthulhu will cache unoptimized frames by hooking into
  `finish(::InferenceState, ::CthulhuInterpreter)` in that case,
  which caches a frame no matter whether it is limited or not.
- when working on a pre-optimization state, if the return type of a
callsite
  is annotated as `LimitedAccuracy`, we widen it and wrap it into
`LimitedCallInfo`.
  When printing `LimitedCallInfo`, Cthulhu only indicates the call is
limited,
  but won't try to print the whole `LimitedAccuracy` object.
- when working on pre-optimization state, we make sure we don't print
`LimitedAccuracy`
  included in `CodeInfo`. Currently I wrote an naive hack to just copy
  the original `CodeInfo` and widen `LimitedAccuracy`s in
`ssavaluetypes` and `rettype`.