Fix handling of MethodResultPure
Formerly Cthulhu tried to ignore statements whose `info` was
`MethodResultPure` . This was marked as an annotation TODO
but also didn't quite work: it turns out this happens because the
compiler introduced a `MethodResultPure` wrapped in a
`UnionSplitApplyCallInfo`, and our logic for bailing out from pure
happened outside of the `map` call used in handling
`UnionSplitApplyCallInfo`.

This fixes it by defining a new `CallInfo` subtype for marking
these `info`s.

Co-authored-by: Shuhei Kadowaki <aviatesk@gmail.com>