Refactor Cthulhu to allow custom callinfo for external packages
This is intended to allow Diffactor to provide a Cthulhu interfaces
for navigating between different differentiation levels. The core
interface is the introduction of `Cthulhu.AbstractCursor`, which
downstream packages can extend to provide additional information
beyond an mi to identify the frame that's currently being displayed.
These interfaces will probably need to evolve a bit, but I think this
is a reasonable start.

Also strengthen the tests a bit to make failures more obvious and
easier to debug.