Add TypedSyntax package
This package attempts to map statements in type-inferred code back
to the source code as written by the programmer.

The intention is to use this in Cthulhu to present the results of
inference in an easier-to-digest form. There are, of course, potential
additional applications of this source-mapping, which is why it is
developed as a semi-independent package.