Stop matching unsupplied args
The function signature is handled specially: we go to effort
to match all arguments, even if they don't have names.
(e.g., `f(::AbstractMatrix)`). Previously, this included arguments
with default values, e.g., if `foo(x, y=2)` then for a call `foo(x)`
we assigned `y::Int`. However, this has two (interwoven) downsides:

- `foo(x)` is actually a different method, and `y` isn't an argument
  for the method we called. So it's a bit weird to give it a type.
- since it is a different method, we can't look up the types in the
  CodeInfo. While we can easily extract the types of literals
  from the source-text, when the default values get more
  complicated, e.g. `y = Int[i^2 for i = 1:3]`, we'd actually have
  run type-inference on that as an expression.

Thus for consistency, restrict type-matching to the arguments we have,
and abandon any attempt to assign types to unsupplied arguments.

While this is a change in behavior, it's analogous to Julia's
type-inference changing its outputs. Consequently I don't view
this as a breaking change. However, let's call this a new "feature"
and bump the TypedSyntax version to 1.1.0.