Fixes for `MethodInstance` objects that do not come from a Method
When running `descend()` on a MethodInstance that does not come from a
`Method`, we need to avoid attempting to tie back to the original source
or method definition names.  This is the case when you have a chunk of
IR that you have generated, but no real method to map back to it with.

I've added some tests showing how you create such a barebones method
instance.