make Cthulhu use Compiler.jl as stdlib optionally when it's loaded (#619)
* make Cthulhu use Compiler.jl as stdlib optionally when it's loaded

Using the Compiler stdlib does make things a bit tricky because we have
to consider Compiler.jl's compatibility not just with the package itself,
but also with the runtime system. In the past, we only had to worry
about the former, so issues didn't come up as often (since using
`Core.Compiler` guarantees that the compiler is compatible with the
runtime system always).

This commit changes Cthulhu back to using `Base.Compiler` until the user
specifically loads the `Compiler.jl` standard library.
This should reduce how often Cthulhu breaks (in normal use).

When an user loads Compiler explicitly, then Cthulhu now load the
CthulhuCompilerExt extension module, and wraps the implementations of
exported APIs.

* temp: don't load Compiler.jl stdlib for now

* rm activation, accept all 0.x.y and 1.x.y versions of Compiler stdlib

* update Project.toml