Fix & clarify IRCode/CodeInfo interactions in LookupResult
`IRCode` is now used primarily for callsite detection and display.
`CodeInfo` may also be used for that purpose, if no `IRCode` is
available. The main requirement is that they are consistent with the
provided call infos.

`CodeInfo` is mainly used to extract slotnames, get more accurate
sources via TypedSyntax, and for module dumping for LLVM/native views.