Split compiler integration to avoid duplicating Cthulhu when loading the compiler extension (#677)
* Initial implementation

* WIP: Overhaul internal architecture & user-facing API

* Bump major version

* Fix highlighter command naming

* Implement initial external AbstractInterpreter support

* Separate commands and their values

This way, commands merely register shortcuts
to execute callbacks, so they get simplified,
and at the same time we no longer have to worry
about keeping command "values" and configuration values
in sync.

* Properly implement way to conditionally disable commands

* Move quit/bookmark commands to beginning of action list

* Fix a few tests

* Adjust conditional Command printing

* `with_effects` -> `effects`, re-enable remark/effect/exception type annotations

The renaming was done to improve consistency with similar settings,
`remarks` and `exception_type`.

* Re-enable VSCode diagnostics & inlay types

* Add backspace shortcut to go back (ascend)

* Refactor bookmark functionality

* Remove disabled AbstractCursor interface

* Fix & clarify IRCode/CodeInfo interactions in LookupResult

`IRCode` is now used primarily for callsite detection and display.
`CodeInfo` may also be used for that purpose, if no `IRCode` is
available. The main requirement is that they are consistent with the
provided call infos.

`CodeInfo` is mainly used to extract slotnames, get more accurate
sources via TypedSyntax, and for module dumping for LLVM/native views.

* Minor interface refactors

* Move terminal utility functions to a separate file

* Start testing various AbstractProvider implementations

* Start adding docs

* Re-enable Compiler extension, test DefaultProvider with both
integrations

* Prevent terminal tests from hanging

* Fix test_terminal.jl

* FakeTerminal -> VirtualTerminal

* Revert minor printing changes

* Fix AbstractInterpreter tests

* Fix test_codeview

* Use unoptimized lookup if source is not available

* More fixes, add a few more tests

* Reimplement `descend_code_`typed/warntype

* Attempt to fix tests for 1.13

* Remove dead code

* Add ExternalProvider example

* Improve interface documentation

* exception_type -> exception_types

To match the plural used for `remarks`, `effects`, etc

* inline_cost -> inlining_costs

To improve consistency with `remarks`, `effects`, `exception_type`

* Update example `descend` usage in README

* Test fixes, avoid corrupting the global method table

* Remove `interruptexc`, add a few more docs

Now that we have backspace to ascend, we can have
'q' quit unconditionally and remove this option.
Furthermore, we now exit without having to handle an interrupt.

* Only reevaluate compiler-related functionality in CthulhuCompilerExt

* Fix tests

* Add more comments, make ir_to_src common

* Mention docstring of AbstractProvider in changelog

* Allow returning `nothing` after lookup to log an error